fix(repositories): change entity ID types from Long to Integer

+++ src/main/java/com/acd/researchrepo/repository/DocumentRequestRepository.java
@@ -4,6 +4,6 @@ import com.acd.researchrepo.model.DocumentRequest;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentRequestRepository extends JpaRepository<DocumentRequest, Long> {
+public interface DocumentRequestRepository extends JpaRepository<DocumentRequest, Integer> {
 
 }
